GL: Implement `glEnd()`
This function performs all the transforms on the current vertex data,
primitive assembly, and writing of the vertex data to the correct
registers (as well as obviously issuing the triangle command).

This currently only supports `GL_TRIANGLES`, but can be easily modified
to support the other primitives that OpenGL supports. There is
currently no clipping, which can result in segfaults from the chipsim
when it tries to draw pixels that are out of bounds.